Refactor distributed query execution into new execution plan
