+++ ballista/rust/core/src/execution_plans/mod.rs
@@ -18,10 +18,12 @@
 //! This module contains execution plans that are needed to distribute Datafusion's execution plans into
 //! several Ballista executors.
 
+mod distributed_query;
 mod shuffle_reader;
 mod shuffle_writer;
 mod unresolved_shuffle;
 
+pub use distributed_query::DistributedQueryExec;
 pub use shuffle_reader::ShuffleReaderExec;
 pub use shuffle_writer::ShuffleWriterExec;
 pub use unresolved_shuffle::UnresolvedShuffleExec;
